+++ src/AppBody.js
@@ -4,7 +4,7 @@ import { Grid } from 'semantic-ui-react'
 export default class AppBody extends Component {
   render() {
     return(
-      <Grid.Row>
+      <Grid.Row color='black'>
         <Grid.Column>
           { this.props.children }
         </Grid.Column>