+++ src/AppFooter.js
@@ -5,7 +5,7 @@ import { Grid } from 'semantic-ui-react'
 export default class AppFooter extends Component {
   render() {
     return(
-      <Grid.Row>
+      <Grid.Row color='black'>
         <Grid.Column>
           { this.props.children }
         </Grid.Column>