+++ src/VehicleGroupStatistics.js
@@ -5,14 +5,14 @@ import VehicleGroupStatistic from './VehicleGroupStatistic'
 export default class VehicleGroupStatistics extends Component {
   render() {
     return(
-      <Grid.Row>
+      <Grid.Row color='black'>
         <Grid.Column width={16}>
-          <Header as='h2' content='Vehicle Info.'/>
+          <Header as='h2' color='blue' content='Vehicle Info.' inverted/>
         </Grid.Column>
         {this.vehicleGroups.map((vehicleGroup) =>
         <Grid.Column computer={5} key={vehicleGroup.key} largeScreen={4}
           tablet={6}>
-          <VehicleGroupStatistic vehicleGroup={vehicleGroup} />
+          <VehicleGroupStatistic vehicleGroup={vehicleGroup}/>
         </Grid.Column>
         )}
       </Grid.Row>)